pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teenyfactories"
version = "0.1.0"
description = "Multi-provider LLM and message queue abstraction for distributed agent systems"
readme = "README.md"
authors = [
    {name = "TeenyFactories Team"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["llm", "agent", "distributed", "message-queue", "openai", "anthropic", "redis", "postgres"]

dependencies = [
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "schedule>=1.1.0",
]

[project.optional-dependencies]
# LLM providers
openai = ["langchain-openai>=0.0.5"]
anthropic = ["langchain-anthropic>=0.1.0"]
google = ["langchain-google-genai>=0.0.5"]
ollama = ["langchain-community>=0.0.20"]
azure = ["langchain-openai>=0.0.5", "openai>=1.0.0"]

# Message queue providers
redis = ["redis>=4.0.0"]
postgres = ["psycopg2-binary>=2.9.0"]

# All LLM providers
all-llm = [
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "langchain-community>=0.0.20",
    "openai>=1.0.0",
]

# All message queue providers
all-mq = [
    "redis>=4.0.0",
    "psycopg2-binary>=2.9.0",
]

# Everything
all = [
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "langchain-community>=0.0.20",
    "openai>=1.0.0",
    "redis>=4.0.0",
    "psycopg2-binary>=2.9.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/teenyfactories/core"
Documentation = "https://github.com/teenyfactories/core/blob/main/README.md"
Repository = "https://github.com/teenyfactories/core"
Issues = "https://github.com/teenyfactories/core/issues"

[tool.setuptools]
packages = ["teenyfactories"]

[tool.setuptools.package-data]
teenyfactories = ["py.typed"]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
